refactor(core): split source files into folders
split v2 vs v4 and allow for helpers regarding core OData attributes or mechanisms.
