+++ packages/odata-core/src/helper/EitherOr.ts
