+++ packages/odata-core/src/index.ts
@@ -1,12 +1,12 @@
 export * from "./ODataVersions";
-export * from "./ODataTypesV4";
-export * from "./ODataTypesV2";
-export * from "./ResponseModelV4";
-export * from "./ResponseModelV401";
-export * from "./FlexibleResponseModelV4";
-export * from "./VersionedResponseModelV4";
-export * from "./ResponseModelV2";
-export * from "./ODataModelPayloadV4";
-export * from "./ODataModelPayloadV401";
-export * from "./FlexibleModelPayloadV4";
-export * from "./VersionedModelPayloadV4";
+export * from "./v4/ODataTypesV4";
+export * from "./v2/ODataTypesV2";
+export * from "./v4/ResponseModelV4";
+export * from "./v4/ResponseModelV401";
+export * from "./v4/FlexibleResponseModelV4";
+export * from "./v4/VersionedResponseModelV4";
+export * from "./v2/ResponseModelV2";
+export * from "./v4/ODataModelPayloadV4";
+export * from "./v4/ODataModelPayloadV401";
+export * from "./v4/FlexibleModelPayloadV4";
+export * from "./v4/VersionedModelPayloadV4";
