+++ packages/odata-core/src/v2/ODataTypesV2.ts
