+++ packages/odata-core/src/v2/ResponseModelV2.ts
