+++ packages/odata-core/src/v4/ResponseModelV4.ts
